Extract PictureVariant out of Picture::Url
The Picture::Url module included transformations to create a variant out of an image. Let's seperate that in it's own class so we can extract url generation into a class.
